move timelapse assembly and directory management from Camera to Timelapse
