+++ bot/main.py
@@ -205,7 +205,7 @@ async def status(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
 
 
 async def check_unfinished_lapses(bot: telegram.Bot) -> None:
-    files = camera_wrap.detect_unfinished_lapses()
+    files = timelapse.detect_unfinished_lapses()
     if not files:
         return
     await bot.send_chat_action(chat_id=config_wrap.secrets.chat_id, action=ChatAction.TYPING)
@@ -733,7 +733,7 @@ async def button_handler(update: Update, context: ContextTypes.DEFAULT_TYPE) -> 
             )
     elif query.data == "cleanup_timelapse_unfinished":
         await context.bot.send_message(chat_id=config_wrap.secrets.chat_id, text="Removing unfinished timelapses data")
-        camera_wrap.cleanup_unfinished_lapses()
+        timelapse.cleanup_unfinished_lapses()
     elif "gcode:" in query.data:
         await ws_helper.execute_ws_gcode_script(query.data.replace("gcode:", ""))
         delete_query = False
